Agregue bd

+++ el_baron_de_la_cerveza/src/database/models/Contact.js
@@ -20,7 +20,7 @@ module.exports = (sequelize, dataTypes) => {
             type: dataTypes.STRING(100)
         },
         userId: {
-            type: dataTypes.INTEGER,
+            type: dataTypes.INTEGER(11),
             allowNull: false
         },
     }
